refactor: move window-stub to utilities
