+++ libs/base/utilities/window-stub/index.ts
